requirements.txt
click==8.1.7
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.4
pillow==10.4.0
instagrapi==2.1.5
python-dotenv==1.0.1
rich==13.7.1
pandas==2.2.2
openpyxl==3.1.2